Backend/src/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
SQLAlchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=0.21
pydantic>=2.0
fastapi
apispec
apispec_webframeworks
flask-swagger-ui